Plain source URLs support

+++ process.go
@@ -59,26 +59,26 @@ func initVips() {
 
 	vipsSupportSmartcrop = C.vips_support_smartcrop() == 1
 
-	if int(C.vips_type_find_load_go(imageTypeJPEG)) != 0 {
+	if int(C.vips_type_find_load_go(C.int(imageTypeJPEG))) != 0 {
 		vipsTypeSupportLoad[imageTypeJPEG] = true
 	}
-	if int(C.vips_type_find_load_go(imageTypePNG)) != 0 {
+	if int(C.vips_type_find_load_go(C.int(imageTypePNG))) != 0 {
 		vipsTypeSupportLoad[imageTypePNG] = true
 	}
-	if int(C.vips_type_find_load_go(imageTypeWEBP)) != 0 {
+	if int(C.vips_type_find_load_go(C.int(imageTypeWEBP))) != 0 {
 		vipsTypeSupportLoad[imageTypeWEBP] = true
 	}
-	if int(C.vips_type_find_load_go(imageTypeGIF)) != 0 {
+	if int(C.vips_type_find_load_go(C.int(imageTypeGIF))) != 0 {
 		vipsTypeSupportLoad[imageTypeGIF] = true
 	}
 
-	if int(C.vips_type_find_save_go(imageTypeJPEG)) != 0 {
+	if int(C.vips_type_find_save_go(C.int(imageTypeJPEG))) != 0 {
 		vipsTypeSupportSave[imageTypeJPEG] = true
 	}
-	if int(C.vips_type_find_save_go(imageTypePNG)) != 0 {
+	if int(C.vips_type_find_save_go(C.int(imageTypePNG))) != 0 {
 		vipsTypeSupportSave[imageTypePNG] = true
 	}
-	if int(C.vips_type_find_save_go(imageTypeWEBP)) != 0 {
+	if int(C.vips_type_find_save_go(C.int(imageTypeWEBP))) != 0 {
 		vipsTypeSupportSave[imageTypeWEBP] = true
 	}
